% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmBound_functions.R
\name{getHarmBound}
\alias{getHarmBound}
\title{Harm boundaries for safety testing}
\usage{
getHarmBound(
  nevents,
  alpha_test = NULL,
  alpha_total = NULL,
  power = NULL,
  pH0,
  maxevents = NULL,
  icc = NULL,
  pH1 = NULL,
  rrH1 = NULL,
  orH1 = NULL,
  rdH1 = NULL,
  r0 = NULL
)
}
\arguments{
\item{nevents}{vector with number of events (over both arms) at which an interim analysis is done}

\item{alpha_test}{the nominal alpha level to use for each test}

\item{alpha_total}{target overall family-wise type I error}

\item{power}{target power at the specified alternative}

\item{pH0}{proportion of events in the treatment arm under the null hypothesis,
typically based on randomization ratio (e.g. 0.5 for a 1:1 randomization)}

\item{maxevents}{optional maximum number of events expected for the trial (over both arms),
used to calculate the expected number of events}

\item{icc}{intraclass correlation if there is more than one event per patient}

\item{pH1}{optional alternative, numeric vector, proportion of events in the treatment arm}

\item{rrH1}{alternative specification of alternative as risk ratio (treatment / control)}

\item{orH1}{alternative specification of alternative as risk ratio (treatment / control). Requires the control proportion (r0).}

\item{rdH1}{alternative specification of alternative as risk difference (treatment - control). Requires the control proportion (r0).}

\item{r0}{risk in the control arm. Required if the alternative is given as risk difference or odds ratio.}
}
\value{
a list with 3 data.frames: bounds, stopprob and opchar.
bounds has a row for each interim analysis and columns for
number of events (events),
number of events in control and treatment arm that would lead to a stop
(events_treatment, events_control), and the nominal alpha for each test (alpha_test).
stopprob has a row for each interim analysis and columns for
number of events (events),
the hypothesis (pH),
the stopping probability (stop_prob), and
the cumulative stopping probability (cum_stop_prob)
opchar has a row for each hypothesis (null plus each alternative) and columns
for the assumed proportion of events in the treatment arm (p),
the cumulative stopping probabilities (cum_stop_prob) and
the expected total number of events (expected_events)
for the null and each alternative.
}
\description{
Calculates the boundaries at each interim analysis, i.e. the number of events in the treatment arm
that would lead to a stopping of the trial based binomial exact tests,
assuming that the events should be equally distributed among both arms.
The indicated scenario (and all more extreme)
would lead to a rejection of H0 (equal distribution) and a stopping for safety.
}
\details{
If there are several events per patient, the intraclass correlation coefficient has to be given
and a beta-binomial framework is used. The overdispersion factor (or design effect)
by which the variance exceeds the regular binomial variance is printed. Note that the
effective sample size is reduced by that factor.

The rejection region for the binomial or beta-binomial exact tests must be given for either
each test (alpha_test), overall (alpha_total, the family-wise error rate) or
by the targetted power for the specified alternative.
}
\examples{

getHarmBound(nevents=c(10,50,100), alpha_test=0.025, pH0=0.5)
#adding an alternative
getHarmBound(nevents=c(10,50,100), alpha_test=0.025, pH0=0.5, pH1=0.6)

#assume that a total of 150 events might occur (for the expected events)
getHarmBound(nevents=c(10,50,100), alpha_test=0.025, pH0=0.5, pH1=0.6, maxevents=150)

#several alternatives
getHarmBound(nevents=c(10,50,100), alpha_test=0.025, pH0=0.5,
pH1 = seq(0.6,0.8,by=0.05), maxevents=150)

#using a risk ratio to specify the alternative
getHarmBound(nevents=c(10,50,100), alpha_test=0.025, pH0=0.5, rrH1=1.5, maxevents=150)

# define the test so that an family-wise type I error of 5\% is achieved
getHarmBound(nevents=c(10,50,100), alpha_total=0.05, pH0=0.5)

# define the test so that an over power of 80\% is achieved
# needs an alternative
getHarmBound(nevents=c(10,50,100), power=0.8, pH0=0.5, pH1=0.6)
}
\seealso{
Available plot methods: \code{\link{plot.harmbound}}
}
