% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmbound_plot.R
\name{opcharNPlot}
\alias{opcharNPlot}
\title{Plot expected number of events}
\usage{
opcharNPlot(harmbound)
}
\arguments{
\item{harmbound}{harmbounds objects as generated using the getHarmBound function}
}
\value{
line plot with the expected number of events for all alternatives
}
\description{
Plot expected number of events
}
\examples{
harmbound<-getHarmBound(nevents=seq(10,100,by=10),alpha_test=0.025,
pH0=0.5,pH1=seq(0.5,0.8,by=0.05),maxevents=150)
opcharNPlot(harmbound)

}
