% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmbound_plot.R
\name{plot.harmbound}
\alias{plot.harmbound}
\title{Plot method for harmbound objects produced by \code{getHarmBound}}
\usage{
\method{plot}{harmbound}(x, ...)
}
\arguments{
\item{x}{harmbounds objects as generated using the getHarmBound function}

\item{...}{further arguments for getHarmBound}
}
\description{
Plot method for harmbound objects produced by \code{getHarmBound}
}
