% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmBound_functions.R
\name{semiConstSpending}
\alias{semiConstSpending}
\title{Title}
\usage{
semiConstSpending(x, alphaVals, startHarmMonitor)
}
\arguments{
\item{x}{the total number of infections (vacc + placebo)}

\item{alphaVals}{is a vector of nominal (un-adjusted) p-value thresholds
to use in establishing cutoffs for harm-monitoring.  This vector
must have the same length as \code{startHarmMonitor}.  The i-th value
of \code{alphaVals} applies to the i-th interval defined by \code{startHarmMonitor}}

\item{startHarmMonitor}{gives the endpoints of all intervals.
The starting point of the first interval is 1, and of the the i-th interval
(for i>1) is 1 + \code{startHarmMonitor}[i-1]}
}
\value{

}
\description{
Title
}
