% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_safety_stop.R
\name{sim_safety_stop}
\alias{sim_safety_stop}
\title{Simulate safety stopping values}
\usage{
sim_safety_stop(nevents, pH0 = 0.5, pH1 = 0.5, alpha_test = 0.025)
}
\arguments{
\item{nevents}{vector with number of events at which an interim analysis is done}

\item{pH0}{proportion of events in the experimental arm under the null hypothesis,
typically based on randomization ratio (e.g. 0.5 for a 1:1 randomization)}

\item{pH1}{proportion of events in the experimental arm under the alternative hypothesis}

\item{alpha_test}{nominal alpha level for binomial exact test}
}
\value{
list with a dataframe with number of events in each group plus upper limit for stopping and indicator for whether stopped, plus indicators number of stops and time points at first stop
}
\description{
Simulate safety stopping values
}
\examples{
set.seed(1)	
sim_safety_stop(nevents=seq(10,100,by=10),pH0 = 0.5, pH1 = 0.7,alpha_test=0.025)


}
