% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmbound_plot.R
\name{absstopPlot}
\alias{absstopPlot}
\title{Plot absolute stopping probs}
\usage{
absstopPlot(harmbound)
}
\arguments{
\item{harmbound}{harmbounds objects as generated using the getHarmBound function}
}
\value{
barplot with stopping probabilities under H0 and optionally H1
}
\description{
Plot absolute stopping probs
}
\examples{
harmbound<-getHarmBound(nevents=seq(10,100,by=10),alpha_test=0.025,pH0=0.5,pH1=c(0.6,0.7,0.8))
absstopPlot(harmbound)

}
