% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmBound_functions.R
\name{convertRisks}
\alias{convertRisks}
\title{Convert the proportion of events in the intervention groups to risk differences and ratios (and vice versa)}
\usage{
convertRisks(
  eprop = NULL,
  etotal = NULL,
  rd = NULL,
  rr = NULL,
  or = NULL,
  r0 = NULL,
  n0,
  n1 = n0
)
}
\arguments{
\item{eprop}{proportion of events in intervention group}

\item{etotal}{total number of events}

\item{rd}{risk difference}

\item{rr}{risk ratio}

\item{or}{odds ratio}

\item{r0}{risk in the control group}

\item{n0}{number of patients in the control group}

\item{n1}{number of patients in the intervention group}
}
\value{
vector with risks in control and intervention group (r0, r1),
the risk difference (rd), risk ratio (rr) and odds ratio (or)
}
\description{
Convert the proportion of events in the intervention groups to risk differences and ratios (and vice versa)
}
\examples{
convertRisks(eprop=0.5,etotal=100,n0=200)
convertRisks(eprop=0.6,etotal=100,n0=200)
convertRisks(rr=1.5,n0=200,r0=0.2)

}
