% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmBound_functions.R
\name{getAlphaPerTest}
\alias{getAlphaPerTest}
\title{Alpha per test}
\usage{
getAlphaPerTest(
  harmMonitorRange,
  null.p,
  totalAlpha = 0.05,
  alpha.interval = c(1e-06, 0.05)
)
}
\arguments{
\item{harmMonitorRange}{harmMonitorRange}

\item{null.p}{null.p}

\item{totalAlpha}{totalAlpha}

\item{alpha.interval}{alpha.interval}
}
\value{

}
\description{
Alpha per test
}
