% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmBound_functions.R
\name{getAlphaPerTest}
\alias{getAlphaPerTest}
\title{Test-wise alpha necessary to control the overall type I error at a specified level (0.05 by default)}
\usage{
getAlphaPerTest(
  nevents,
  totalAlpha = 0.05,
  pH0 = 0.5,
  alpha.interval = c(10^(-10), 0.05)
)
}
\arguments{
\item{nevents}{vector with number of events at which an interim analysis is done}

\item{totalAlpha}{Overall type I error, 0.05 by default}

\item{pH0}{proportion of events in the intervention arm under the null hypothesis,
typically based on randomization ratio (e.g. 0.5 for a 1:1 randomization)}

\item{alpha.interval}{Range for test-wise alpha, c(10^(-10),0.05) by default}
}
\value{
Test-wide alpha
}
\description{
Test-wise alpha necessary to control the overall type I error at a specified level (0.05 by default)
}
\examples{
apt<-getAlphaPerTest(nevents = c(10,50,100), totalAlpha = 0.05, pH0 = 0.5)
apt
getHarmBound(nevents = c(10,50,100),alpha_test = apt, pH0 = 0.5)


}
