% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmBound_functions.R
\name{getHarmBound}
\alias{getHarmBound}
\title{Harm boundaries for safety testing}
\usage{
getHarmBound(nevents, alpha_test, pH0)
}
\arguments{
\item{nevents}{vector with number of events at which an interim analysis is done}

\item{alpha_test}{the nominal alpha level to use for each test}

\item{pH0}{proportion of events in the experimental arm under the null hypothesis,
typically based on randomization ratio (e.g. 0.5 for a 1:1 randomization)}
}
\value{
data frame with a row for each interim analysis and with columns for
number of events (n),
number of events in control and experimental group that would lead to a stop (n_treat, n_control),
risk ratio (RR) at that level,
sopping probability (stopProb),
cumulative stopping probability (cumStopProb).
and the nominal alpha for each test (alphaVal)
}
\description{
Calculates the boundaries at each interim analysis, i.e. the number of events in the experimental group
that would lead to a stopping of the trial based on a binomial exact test,
assuming that the events should be equally distributed amont both groups.
The indicated scenario (and all more extreme) would lead to a rejection of H0 (equal distribution) and a stopping for safety.
}
\examples{
getHarmBound(nevents=c(10,50,100), alpha_test=0.025, pH0=0.5)
}
