% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmbound_plot.R
\name{harmboundPlot}
\alias{harmboundPlot}
\title{Plot harmbounds and the observed events}
\usage{
harmboundPlot(
  harmbound,
  observed = NULL,
  colourbound = c("blue", "red"),
  fill_alpha = 0.5,
  colourobserved = "black",
  H0line = TRUE
)
}
\arguments{
\item{harmbound}{harmbounds objects as generated using the getHarmBound function}

\item{observed}{optional observed number of events,
as a vector with the sequential groups in which an event occured
(0 for control and 1 for experimental)}

\item{colourbound}{vector with two colours for the bounds, in and out, default is blue and red}

\item{fill_alpha}{opacity of the colours for the bounds}

\item{colourobserved}{colour for the line with the observed events}

\item{H0line}{logical, whether a line to indicate the expectations should be added.}
}
\value{
plot with the bounds and optionally the observed number of events
}
\description{
Plot harmbounds and the observed events
}
\examples{
hb<-getHarmBound(nevents=seq(10,100,by=10),alpha_test=0.025,pH0=0.5)
harmboundPlot(hb)
set.seed(123)
eventgroups<-rbinom(n=100,size=1,prob=0.5)
harmboundPlot(hb,observed=eventgroups)

}
