% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmbound_plot.R
\name{opcharStopPlot}
\alias{opcharStopPlot}
\title{Plot cumulative stopping probability by hypothesis}
\usage{
opcharStopPlot(harmbound)
}
\arguments{
\item{harmbound}{harmbounds objects as generated using the getHarmBound function}
}
\value{
line plot with the cumulative stopping for all alternatives
}
\description{
Plot cumulative stopping probability by hypothesis
}
\examples{
harmbound<-getHarmBound(nevents=seq(10,100,by=10),alpha_test=0.025,
pH0=0.5,pH1=seq(0.5,0.8,by=0.05),maxevents=150)
opcharStopPlot(harmbound)

}
