% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_safety_stop.R
\name{simSafetyStop}
\alias{simSafetyStop}
\title{Simulate safety stopping}
\usage{
simSafetyStop(
  nevents,
  pH0 = 0.5,
  alpha_test = 0.025,
  pH1 = NULL,
  rrH1 = NULL,
  orH1 = NULL,
  rdH1 = NULL,
  r0 = NULL,
  n = NULL
)
}
\arguments{
\item{nevents}{vector with number of events at which an interim analysis is done}

\item{pH0}{proportion of events in the experimental arm under the null hypothesis,
typically based on randomization ratio (e.g. 0.5 for a 1:1 randomization)}

\item{alpha_test}{nominal alpha level for binomial exact test}

\item{pH1}{proportion of events in the experimental arm under the alternative hypothesis}

\item{rrH1}{risk ratio (experimental / control).}

\item{orH1}{risk ratio (experimental / control). Requires the control proportion (r0).}

\item{rdH1}{risk difference (experimental - control). Requires the control proportion (r0) and the number of participants (n).}

\item{r0}{risk in the control arm. Required if the effect is given as risk difference or odds ratio.}

\item{n}{total number of participants. Required if the effect is given as risk difference.}
}
\value{
list with a dataframe with number of events in each group plus upper limit for stopping and indicator for whether stopped, plus indicators number of stops and time points at first stop
}
\description{
The effect can be given as proportion of events in the experimental group (pH1),
the risk difference (rdH1), risk ratio (rrH1) or odds ratio (orH1).
}
\examples{
set.seed(1)	
simSafetyStop(nevents=seq(10,100,by=10),pH0 = 0.5, pH1 = 0.6,alpha_test=0.025)

set.seed(1)	
simSafetyStop(nevents=seq(10,100,by=10),pH0 = 0.5, rrH1 = 0.6/(1-0.6), alpha_test=0.025)


}
